[MOB] ADD: Directory letter_view, ADD: implemented Widget letter_icon #8
