--- lib/widgets/home/letter_icon_box.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LetterIconBox extends StatefulWidget {
-  const LetterIconBox({super.key});
-
-  @override
-  State<LetterIconBox> createState() => _LetterIconBoxState();
-}
-
-class _LetterIconBoxState extends State<LetterIconBox> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
