+++ lib/widgets/home/letter_view/letter_grid.dart
@@ -29,27 +29,6 @@ class _LetterCardState extends State<LetterGrid> {
 
   @override
   Widget build(BuildContext context) {
-    final titleStyle = TextStyle(
-      fontFamily: 'Inter',
-      fontSize: 20,
-      fontWeight: FontWeight.w400,
-      letterSpacing: -0.5,
-    );
-
-    final dDayStyle = TextStyle(
-      fontFamily: 'Inter',
-      fontSize: 18,
-      fontWeight: FontWeight.w600,
-      letterSpacing: 1,
-    );
-
-    final dateStyle = TextStyle(
-      fontFamily: 'Inter',
-      fontSize: 16,
-      fontWeight: FontWeight.normal,
-      letterSpacing: -0.1,
-      color: Color(0xff6B7280),
-    );
     return SliverPadding(
       padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 0),
       sliver: SliverGrid(
@@ -57,7 +36,7 @@ class _LetterCardState extends State<LetterGrid> {
           crossAxisCount: 2,
           mainAxisSpacing: 5,
           crossAxisSpacing: 5,
-          childAspectRatio: 0.8,
+          childAspectRatio: 0.84,
         ),
         delegate: SliverChildBuilderDelegate(
           (context, index) {
