+++ lib/widgets/home/letter_view/letter_info_box.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class LetterInfoBox extends StatelessWidget {
+  const LetterInfoBox({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
